Поправил замечания наставника
