+++ src/mock/trip-event-point.js
@@ -123,50 +123,50 @@ const mockEvents = [
     id: 1,
     type: getRandomArrayElement(EVENT_POINTS_TYPES),
     destination: getRandomArrayElement(DESTINATION_POINTS).id,
-    date_from: new Date('2019-08-17 09:00'),
-    date_to: new Date('2019-08-17 18:00'),
-    base_price: getRandomInteger(),
-    is_favorite: false,
+    'date_from': new Date('2019-08-17 09:00'),
+    'date_to': new Date('2019-08-17 18:00'),
+    'base_price': getRandomInteger(),
+    'is_favorite': false,
     offers: [ getRandomArrayElement(getRandomArrayElement(OFFERS).offers).id ],
   },
   {
     id: 2,
     type: getRandomArrayElement(EVENT_POINTS_TYPES),
     destination: getRandomArrayElement(DESTINATION_POINTS).id,
-    date_from: new Date('2019-06-05 09:45'),
-    date_to: new Date('2020-01-13 21:30'),
-    base_price: getRandomInteger(),
-    is_favorite: false,
+    'date_from': new Date('2019-06-05 09:45'),
+    'date_to': new Date('2020-01-13 21:30'),
+    'base_price': getRandomInteger(),
+    'is_favorite': false,
     offers: [ getRandomArrayElement(getRandomArrayElement(OFFERS).offers).id ],
   },
   {
     id: 3,
     type: getRandomArrayElement(EVENT_POINTS_TYPES),
     destination: getRandomArrayElement(DESTINATION_POINTS).id,
-    date_from: new Date('2020-02-19 08:20'),
-    date_to: new Date('2020-03-05 02:15'),
-    base_price: getRandomInteger(),
-    is_favorite: true,
+    'date_from': new Date('2020-02-19 08:20'),
+    'date_to': new Date('2020-03-05 02:15'),
+    'base_price': getRandomInteger(),
+    'is_favorite': true,
     offers: [ getRandomArrayElement(getRandomArrayElement(OFFERS).offers).id ],
   },
   {
     id: 4,
     type: getRandomArrayElement(EVENT_POINTS_TYPES),
     destination: getRandomArrayElement(DESTINATION_POINTS).id,
-    date_from: new Date('2020-04-11 09:00'),
-    date_to: new Date('2020-05-17 18:00'),
-    base_price: getRandomInteger(),
-    is_favorite: false,
+    'date_from': new Date('2020-04-11 09:00'),
+    'date_to': new Date('2020-05-17 18:00'),
+    'base_price': getRandomInteger(),
+    'is_favorite': false,
     offers: [ getRandomArrayElement(getRandomArrayElement(OFFERS).offers).id ],
   },
   {
     id: 5,
     type: getRandomArrayElement(EVENT_POINTS_TYPES),
     destination: getRandomArrayElement(DESTINATION_POINTS).id,
-    date_from: new Date('2021-02-14 12:20'),
-    date_to: new Date('2021-04-10 23:30'),
-    base_price: getRandomInteger(),
-    is_favorite: true,
+    'date_from': new Date('2021-02-14 12:20'),
+    'date_to': new Date('2021-04-10 23:30'),
+    'base_price': getRandomInteger(),
+    'is_favorite': true,
     offers: [ getRandomArrayElement(getRandomArrayElement(OFFERS).offers).id ],
   },
 ];
